Update sidebar.lp
Just a suggestion - change the background colour of the sidebar total number of domains from yellow to green. This matches the upper box where it is green, as I keep seeing yellow in the sidebar and thinking it's something to do with blocking like the upper yellow box. It also gives the small white text a little more contrast and makes it more readable.

Signed-off-by: Chris L <github@holosys.co.uk>